requirements.txt
Django==2.1
djangorestframework==3.9.1
Pillow==5.4.1
pkg-resources==0.0.0
pytz==2018.9
